Quick hack to (show usage of /) use cwds/benchmark.
Since this uses asm() it probably only works with
gcc and clang, and certainly only works on intel 64-bit
too.

The benchmark code is mostly intended to measure a
relative speed with a (very) high reproducibility,
so that the resulting values can be compared with
other code. However, it only measures the time
needed with a warmed up cache; which might be
completely different from Real World code when
that code is only executed sporadically.

Moreover, due to the complexity of a CPU the exact
same assembly instructions can take a different
number of clock cycles upon different executions,
depending on the direct history of the CPU;
so, the returned value (although reproducible) are
merely an indication of *roughly* the number of
clock cycles needed (with a hot cache).

The actual spread can (in a statistical sense) can
be much wider and irradic. There is certainly not
an underlaying normal distribution for the number
of clock cycles for a particular piece of code; merely
a number of 'history determined CPU states' each
with a certain probability given the testsuite code
that is executing this. The most disturbing things,
like an interrupt or context switch due to the
schedular has been completely removed from the picture
however, because those events cause such enormous much
longer times (milliseconds instead of nanoseconds)
that they are detectable in 100% of the cases.